Fix auth compatibility with ROS 6.43+
In RouterOS version 6.43 a new form of plain authenticated was introduced.
In RouterOS version 6.45.1 the old authentication method as removed so we now need to use the form introduced in 6.43
see: https://wiki.mikrotik.com/wiki/Manual:API#Initial_login

This changes the steps done to login in the API becoming compatible with the new versions and incompatible with the old ones.

Please beware this library is still using plaintext authentication so isn't secure. Ill try to add HTTPS support when possible ...